+++ src/strauss/presets/__init__.py
@@ -1,3 +1,4 @@
 from . import synth
 from . import sampler
 from . import spec
+from . import speech